apps/consumer-python/requirements-dev.txt
-r requirements.txt
pytest>=8.0
pact-python-cli>=2.0.0
